requirements.txt
agate==1.9.1
altair==6.0.0
annotated-types==0.7.0
asn1crypto==1.5.1
asttokens==3.0.1
attrs==25.4.0
babel==2.17.0
blinker==1.9.0
boto3==1.42.10
botocore==1.42.10
cachetools==6.2.4
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
comm==0.2.3
contourpy==1.3.3
cryptography==46.0.0
cycler==0.12.1
daff==1.4.2
dbt-adapters==1.22.0
dbt-common==1.37.2
dbt-core==1.11.0rc3
dbt-extractor==0.6.0
dbt-protos==1.0.410
dbt-semantic-interfaces==0.9.0
dbt-snowflake==1.10.4
debugpy==1.8.19
decorator==5.2.1
deepdiff==8.6.1
executing==2.2.1
filelock==3.20.1
fonttools==4.61.1
gitdb==4.0.12
GitPython==3.1.45
idna==3.11
importlib_metadata==8.7.0
ipykernel==7.1.0
ipython==9.8.0
ipython_pygments_lexers==1.1.1
isodate==0.7.2
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.3.0
jedi==0.19.2
jeepney==0.9.0
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.3
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.7.0
jupyter_core==5.9.1
keyring==25.7.0
kiwisolver==1.4.9
leather==0.4.1
MarkupSafe==3.0.3
mashumaro==3.14
matplotlib==3.10.8
matplotlib-inline==0.2.1
more-itertools==10.8.0
msgpack==1.1.2
narwhals==2.14.0
nest-asyncio==1.6.0
networkx==3.6.1
numpy==2.3.5
nvidia-nccl-cu12==2.28.9
orderly-set==5.5.0
packaging==25.0
pandas==2.3.3
parsedatetime==2.6
parso==0.8.5
pathspec==0.12.1
pexpect==4.9.0
pillow==12.0.0
platformdirs==4.5.1
plotly==6.5.0
prompt_toolkit==3.0.52
protobuf==6.33.2
psutil==7.1.3
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==22.0.0
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
PyJWT==2.10.1
pyOpenSSL==25.3.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-slugify==8.0.4
pytimeparse==1.1.8
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
s3transfer==0.16.0
scikit-learn==1.8.0
scipy==1.16.3
seaborn==0.13.2
SecretStorage==3.5.0
six==1.17.0
smmap==5.0.2
snowflake-connector-python==3.18.0
snowplow-tracker==1.1.0
sortedcontainers==2.4.0
sqlparse==0.5.4
stack-data==0.6.3
streamlit==1.52.1
tenacity==9.1.2
text-unidecode==1.3
threadpoolctl==3.6.0
toml==0.10.2
tomlkit==0.13.3
tornado==6.5.4
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
watchdog==6.0.0
wcwidth==0.2.14
xgboost==3.1.2
zipp==3.23.0
